Version 1.1.0

+++ conftest.py
@@ -10,5 +10,9 @@ def create_user():
     user = data.generate_user()
     response = requests.post(urls.REGISTER, json=user)
     json = response.json()
-    yield user, response.status_code, json
-    requests.delete(urls.USER, headers={'authorization': json['accessToken']})
+    if response.status_code != 200:
+        raise RuntimeError(f"user creation failure {response.status_code}")
+    yield {'user': user, 'code': response.status_code, 'json': json}
+    response = requests.delete(urls.USER, headers={'authorization': json['accessToken']})
+    if response.status_code != 202:
+        raise RuntimeError(f"user deletion failure {response.status_code}")
